requirements.txt
# 기본 데이터 처리 및 지도 생성
pandas>=1.5.0
folium>=0.20.0

# 웹 요청 및 데이터 다운로드
requests>=2.31.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 지리 정보 처리
geopy>=2.4.0

# 이미지 처리
Pillow>=10.0.0

# 추가 유틸리티 (선택사항)
numpy>=1.24.0
matplotlib>=3.7.0